feat: parse line numbers
When trying to open a link to a file that also has a line number, e.g.
`path/to/file:42` these changes now allow to open the file and put the
cursor on the corresponding line number.

Previously, open a file with a `:` and the line number would fail.

Additionally these changes use `vim.fn.filereadable` instead of `fs_open`/`fs_stat`, as the API
is easier to use and (I think) it may be less costly than opening the
file via `fs_open`.