Resolve links without .md extension again (fix #26)
This commit simplifies the parsing of local links with line numbers and
especially fixes that links without extension do not resolve.

The previous version tried to concatenate modified_link and ".md", but
as modified_link is defined as nil this fails. The solution is to
concatenate with the variable link. But if a line number is specified
this would append ".md" after the line number and the line parsing code
is never reached, e. g.:

  /path/to/file:23.md

Therefore the splitting of path and line number is now executed at the
beginning of the function. This ensures that the path variable does not
contain the line number. For the line_number it is utilized that Lua
defaults to a nil value when a non-existing table index is given.